% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distmatrix.R
\name{distmatrix}
\alias{distmatrix}
\alias{bind_images}
\alias{compute_matrix}
\alias{create_miniature}
\title{Computing image distance matrix.}
\usage{
bind_images(...)

compute_matrix(images)

create_miniature(x, unique_id)
}
\description{
Computing image distance matrix.
}
\examples{
\dontrun{
  imgs <- bind_images('path.to.png', repository(filesystem('path/to/repository')))
  dist <- compute_matrix(imgs)
  path <- map_chr(imgs, create_miniature)
  rownames(dist) <- colnames(dist) <- sprintf("![](\%s)", path)
}
}
