% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{edit-dist}
\alias{edit-dist}
\alias{tokenize}
\alias{edit_dist}
\title{Calculate edit distance between R expressions.}
\usage{
tokenize(x)

edit_dist(a, b)
}
\arguments{
\item{x}{R \link[base:expression]{base::expression} or R expression string (see \link[base:deparse]{base::deparse}).}

\item{a}{tokenized R expression; output of \code{tokenize()}.}

\item{b}{tokenized R expression; output of \code{tokenize()}.}
}
\description{
Calculates edit distance at the level of R tokens. See
\link[sourcetools:tokenize_string]{sourcetools::tokenize_string} for details on tokenizer.
}
\examples{
\dontrun{
edit_dist(tokenize(bquote(x <- 1)), tokenize(bquote(y <- 1)))
}
}
