% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify-artifacts}
\alias{identify-artifacts}
\alias{identify_file}
\alias{identify_object}
\alias{identify_expression}
\title{Identify artifacts by a variety of means.}
\usage{
identify_file(path, repo)

identify_object(obj, repo)

identify_expression(expr, repo, n = 3)
}
\arguments{
\item{path}{file path.}

\item{repo}{a \link[repository:repository]{repository::repository} object.}

\item{obj}{any R object.}

\item{expr}{expression to match against; as returned by \link[base:bquote]{base::bquote} or
\link[base:deparse]{base::deparse}.}

\item{n}{return that many closest matches given the edit distance (see
\link{edit_dist}).}
}
\value{
Each function returns a \code{list} of artifacts (see
\link[repository:new_artifact]{repository::new_artifact}) wrapped in a \link[utilities:as_container]{utilities::as_container}.
}
\description{
\code{identify_file} finds artifacts matching the contents of
a given file: plot, data set (\code{csv}) or an R object (\code{rds} or \code{RData}).

\code{identify_object} finds artifacts matching a given R object.

\code{identify_expression} finds artifacts created with an
expression matching the one passed as argument. Each artifact is
assigned an additional attribute, \code{dist}, which holds the edit
distance to the expression \code{expr}.
}
