% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_file}
\alias{identify_file}
\title{Identify artifact(s) matching a given file.}
\usage{
identify_file(path, repo)
}
\arguments{
\item{path}{file path.}

\item{repo}{a \link[repository:repository]{repository::repository} object.}
}
\value{
artifact(s) matching contents of \code{path}.
}
\description{
Identify artifact(s) matching a given file.
}
\seealso{
identify_object
}
