% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_object}
\alias{identify_object}
\title{Identify artifact(s) matching a given R object.}
\usage{
identify_object(obj, repo)
}
\arguments{
\item{obj}{any R object.}

\item{repo}{a \link[repository:repository]{repository::repository} object.}
}
\value{
artifact(s) matching \code{obj}.
}
\description{
Identify artifact(s) matching a given R object.
}
\seealso{
identify_file
}
