% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{image_dist}
\alias{image_dist}
\title{Compute distance between two images.}
\usage{
image_dist(a, b, cutoff = 0.5)
}
\arguments{
\item{a}{First image; see \link[imager:cimg]{imager::cimg}.}

\item{b}{Second image.}

\item{cutoff}{The cut-off quantile for gradient values.}
}
\value{
Cumulative area between _ecdf_s for all pairs of columns.
}
\description{
Important: images are compared column-wise, thus they must have equal
numbers of columns. Distance is computed in two phases: preparation
followed by computing the proper distance.
}
\details{
During the preparation, each image is first turned into scale of
grays with \link[imager:grayscale]{imager::grayscale}. Then, gradient of color changes is
computed for each image as \code{imgradient(i, 'xy')} (see \link[images:imgradient]{images::imgradient}).
Only gradients larger than \code{quantile(gradient, cutoff)} are kept.

In the second phase, images are matched column-wise, creating pairs
of columns. For each column the distances from the bottom of the image
(bottom row) to each non-zero gradient value is calculated. These
values are then treated as results of sampling a random variable.
The empirical cumulative distribution function (\emph{ecdf}) is obtained
for each column and the area between matching _ecdf_s is calculated.
This is similar to how the Kolmogorov-Smirnov test is performed.

Finally, the total area for all pairs of columns is returned as the
distance between images \code{a} and \code{b}.

For optimal results images should be always \emph{unwrapped}; see
\link{unwrap_image}.
}
\examples{
\dontrun{
  exp <- load.image('roc.png')
  img <- load.image('test.png')
  image_dist(unwrap_image(test), unwrap_image(exp))
}
}
