% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflection.R
\name{reflection}
\alias{reflection}
\title{Reflect the History of an Artifact}
\description{
Matches the contents of a file against a repository of
artifacts. Provides the context in which the while was created.
}
