% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{unwrap_image}
\alias{unwrap_image}
\title{Unwrap an image.}
\usage{
unwrap_image(x, dAlpha = 1, dR = 1, missing = 1)
}
\arguments{
\item{x}{Input image.}

\item{dAlpha}{Step for the angle \code{alpha}.}

\item{dR}{Step for the distance \code{r}.}

\item{missing}{The value for pixels outside of \code{x}.}
}
\description{
Unwrapping means transforming image from polar coordinates into
carthsian coordinates. First, the center of the polar coordinate
system is placed at the center of the image rectangle. Then, pixel
values are read for each angle \code{alpha} from \code{0} to \code{2*Pi}, with a
step of \code{dAlpha}, and distance \code{r} from \code{0} to \code{diag/2}, with a step
of \code{dR}. Each sun "ray" becomes a column of the output image.
}
\details{
Because maximum distance \code{diag/2} defines a circle extending beyond
the input image \code{x}, pixel values outside of \code{x} are assumed to be
equal to \code{missing}.
}
