% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/annotate.R
\name{edit_dist}
\alias{edit_dist}
\alias{edit-dist}
\alias{tokenize}
\title{Calculate edit distance between R expressions.}
\usage{
edit_dist(a, b)

tokenize(x)
}
\arguments{
\item{a}{tokenized R expression; output of \code{tokenize()}.}

\item{b}{tokenized R expression; output of \code{tokenize()}.}

\item{x}{R \link[base:expression]{base::expression} or R expression string (see \link[base:deparse]{base::deparse}).}
}
\description{
Calculates edit distance at the level of R tokens. See
\link[sourcetools:tokenize_string]{sourcetools::tokenize_string} for more information about the
tokenizer.
}
\examples{
\dontrun{
edit_dist(tokenize(bquote(x <- 1)), tokenize(bquote(y <- 1)))
}
}
